test: Test for index errors and replication.
